+++ compare-frontend/src/app/modules/map/converter.service.ts
@@ -35,8 +35,6 @@ export const converterToken = new InjectionToken<Converter>('Converter');
 export interface LayerComposite {
     id: string,
     layer: Layer
-    // legend: { component: Type<any>, args: {[key: string]: any} }
-    // info: { component: Type<any>, args: {[key: string]: any} }
     popup(location: number[], features: FeatureLike[]): { component: Type<any>, args: {[key: string]: any} } | undefined
     onClick(location: number[], features: FeatureLike[]): void
     onHover(location: number[], features: FeatureLike[]): void