+++ compare-frontend/src/app/modules/wizard/wizard.service.ts
@@ -3,7 +3,6 @@ import { Store } from "@ngrx/store";
 import { Observable, OperatorFunction, filter, map } from "rxjs";
 import { Partition, RiesgosProduct, RiesgosScenarioState, RiesgosState, RiesgosStep, ScenarioName } from "src/app/state/state";
 import * as AppActions from 'src/app/state/actions';
-import { WizardModule } from "./wizard.module";
 
 
 export interface StepData {